+++ R/rmweather_plotting_functions.R
@@ -30,6 +30,8 @@ rmw_plot_partial_dependencies <- function(df) {
 #' 
 #' @param df Data frame created by \code{\link{rmw_model_importance}}. 
 #' 
+#' @param colour Colour of point and segment geometries.
+#' 
 #' @author Stuart K. Grange
 #' 
 #' @return ggplot2 plot with point and segment geometries.
@@ -37,7 +39,7 @@ rmw_plot_partial_dependencies <- function(df) {
 #' @seealso \code{\link{rmw_train_model}}, \code{\link{rmw_model_importance}}
 #' 
 #' @export
-rmw_plot_importance <- function(df) {
+rmw_plot_importance <- function(df, colour = "black") {
   
   # Check input
   if (!all(c("rank", "variable", "importance") %in% names(df))) {
@@ -54,12 +56,14 @@ rmw_plot_importance <- function(df) {
     df, 
     ggplot2::aes(importance, reorder(variable, importance))
   ) + 
-    ggplot2::geom_point(size = 5) + 
+    ggplot2::geom_point(size = 5, colour = colour) + 
     ggplot2::geom_segment(
       ggplot2::aes(
         x = 0, y = reorder(variable, importance), 
         xend = importance, 
-        yend = reorder(variable, importance))
+        yend = reorder(variable, importance)
+      ),
+      colour = colour
     ) +
     ggplot2::theme_minimal() + 
     ggplot2::ylab("Variable") + 
@@ -101,7 +105,9 @@ rmw_plot_test_prediction <- function(df, bins = 30) {
       option = "inferno",
       begin = 0.3,
       end = 0.8
-    )
+    ) + 
+    ggplot2::xlab("Observed") + 
+    ggplot2::ylab("Predicted")
   
   return(plot)
   
@@ -111,13 +117,16 @@ rmw_plot_test_prediction <- function(df, bins = 30) {
 #' Function to plot the meteorologically normalised time series after
 #' \code{\link{rmw_normalise}}. 
 #' 
+#' If the input data contains a standard error variable named \code{"se"}, 
+#' this will be plotted as a ribbon (+ and -) around the mean. 
+#' 
 #' @param df Data frame created by \code{\link{rmw_normalise}}. 
 #' 
 #' @param colour Colour for line geometry. 
 #' 
 #' @author Stuart K. Grange
 #' 
-#' @return ggplot2 plot with a line geometry. 
+#' @return ggplot2 plot with a line and ribbon geometries. 
 #' 
 #' @examples 
 #' 
@@ -127,8 +136,27 @@ rmw_plot_test_prediction <- function(df, bins = 30) {
 #' @export
 rmw_plot_normalised <- function(df, colour = "#6B186EFF") {
   
-  plot <- ggplot2::ggplot() + 
-    ggplot2::geom_line(data = df, ggplot2::aes(date, value_predict), colour = colour) + 
+  # Create plot
+  plot <- ggplot2::ggplot(data = df)
+  
+  # Add se if present in data frame too
+  if ("se" %in% names(df)) {
+    
+    # Plot the ribbon geom first
+    plot <- plot + 
+      ggplot2::geom_ribbon(
+        ggplot2::aes(x = date, ymin = value_predict - se, ymax = value_predict + se), 
+        alpha = 0.3,
+        fill = "grey"
+      )
+    
+  }
+  
+  # Overlay line  
+  plot <- plot + 
+    ggplot2::geom_line(
+      ggplot2::aes(date, value_predict), colour = colour
+    ) + 
     ggplot2::theme_minimal() +
     ggplot2::ylab("Meteorologically normalised value") + 
     ggplot2::xlab("Date")